Update struktur database ke migration laravel

+++ database/migrations/2025_04_21_103028_update_transaksi_p_kehamilan_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
         // Buat ulang tabel
         Schema::create('transaksi_p_kehamilan', function (Blueprint $table) {
             $table->id();
-            $table->integer('id_pendaftaran');
+            $table->unsignedBigInteger('id_pendaftaran');
 
             $table->string('pendampingan')->nullable();
             $table->text('keluhan_utama')->nullable();
