+++ database/migrations/2025_04_21_234601_update_transaksi_p_anak_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
 
         Schema::create('transaksi_p_anak', function (Blueprint $table) {
             $table->id();
-            $table->integer('id_pendaftaran');
+            $table->unsignedBigInteger('id_pendaftaran');
 
             $table->string('nama')->nullable();
             $table->date('tanggal_lahir')->nullable();
